weather_forcast/requirement.txt
uvicorn
fastapi
httpx 
cachetools 
python-dotenv